// web/src/api/generated/model/workflowDefSchemaInput.ts
/**
 * Generated by orval v8.15.0 🍺
 * Do not edit manually.
 * Mixd
 * Personal music metadata hub
 * OpenAPI spec version: 0.8.9
 */
import type { WorkflowTaskDefSchemaInput } from './workflowTaskDefSchemaInput.ts';

export interface WorkflowDefSchemaInput {
  id: string;
  name: string;
  description?: string;
  version?: string;
  tasks?: WorkflowTaskDefSchemaInput[];
}